Keep the engine type surface off the hooks layer
Move the RenderingExtensions interface into interfaces/ so
QItemOverrideComponentProps no longer types against
hooks/useRenderingExtensions, whose module graph reaches
html-react-parser, and import UseResponsiveProps from
hooks/useResponsive directly instead of through the hooks barrel.
Both edges pulled UI packages into the type closure of the headless
entrypoint that only one deliberate MUI breakpoint type should reach.
